+++ package.json
@@ -9,49 +9,40 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@chakra-ui/anatomy": "^2.2.1",
-    "@chakra-ui/icons": "^2.1.1",
-    "@chakra-ui/react": "^2.8.1",
+    "@chakra-ui/react": "^2.8.2",
     "@emotion/react": "^11.11.0",
     "@emotion/styled": "^11.11.0",
     "@fontsource-variable/noto-sans-mono": "^5.0.18",
     "@formkit/auto-animate": "1.0.0-beta.6",
     "@legendapp/state": "^2.1.4",
     "@reduxjs/toolkit": "^1.9.5",
-    "@types/randomcolor": "^0.5.9",
     "@types/react": "^18.2.7",
     "@types/react-dom": "^18.2.4",
-    "@use-gesture/react": "^10.3.0",
-    "dough-js": "^0.0.8",
-    "emoji-from-text": "^1.1.12",
     "framer-motion": "^10.12.16",
-    "ramda": "^0.29.0",
+    "ramda": "^0.29.1",
     "rand-seed": "^1.0.2",
     "randomcolor": "^0.6.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.2",
-    "react-swipeable": "^7.0.1",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.2",
-    "swiper": "^11.0.4",
-    "tua-body-scroll-lock": "^1.4.0",
     "use-long-press": "^3.2.0"
   },
   "devDependencies": {
+    "@types/randomcolor": "^0.5.9",
     "@types/ramda": "^0.29.2",
     "autoprefixer": "^10.4.14",
     "postcss": "^8.4.24",
     "prettier": "2.8.8",
-    "rome": "12.1.3",
     "typescript": "^5.0.4",
-    "vite": "^4.3.9",
-    "vite-plugin-pwa": "^0.15.1",
-    "workbox-build": "^6.6.0",
-    "workbox-core": "^6.6.0",
-    "workbox-precaching": "^6.6.0",
-    "workbox-routing": "^6.6.0",
-    "workbox-strategies": "^6.6.0",
-    "workbox-window": "^6.6.0"
+    "vite": "^5.0.12",
+    "vite-plugin-pwa": "^0.17.4",
+    "workbox-build": "^7.0.0",
+    "workbox-core": "^7.0.0",
+    "workbox-precaching": "^7.0.0",
+    "workbox-routing": "^7.0.0",
+    "workbox-strategies": "^7.0.0",
+    "workbox-window": "^7.0.0"
   }
 }
