+++ src/Adder.tsx
@@ -5,24 +5,22 @@ import {
   InputLeftElement,
   forwardRef,
 } from "@chakra-ui/react";
-import { ChangeEventHandler, useContext, useEffect, useState } from "react";
+import { ChangeEventHandler, useContext, useState } from "react";
 
 import * as R from "ramda";
-import Rand from "rand-seed";
 import { CoordinatesContext } from "./App";
 import { Plusik } from "./Plusik";
-import { emojis } from "./emojis";
+import { getRandomEmoji } from "./emojis";
 import {
+  TodoState,
   addTask,
   renameScreen,
   useAppDispatch,
   useAppSelector,
   type Todo,
-  TodoState,
 } from "./store";
 export interface Props extends InputGroupProps {
   what: "task" | "screen";
-  taskMode?: "fast" | "slow";
   initialEmoji?: string;
   where?: keyof TodoState;
 }
@@ -32,7 +30,6 @@ const Adder = forwardRef<Props, "div">((props, ref) => {
     placeholder,
     what = "task",
     initialEmoji = "+",
-    taskMode,
     where,
     ...inputGroupProps
   } = props;
@@ -40,7 +37,6 @@ const Adder = forwardRef<Props, "div">((props, ref) => {
   const [row, column] = useContext(CoordinatesContext);
   const structure = useAppSelector((state) => state.todo.structure);
 
-  const [mode, setMode] = useState(taskMode);
   const [text, setText] = useState("");
 
   const handleChange: ChangeEventHandler<HTMLInputElement> = R.pipe(
@@ -86,11 +82,19 @@ const Adder = forwardRef<Props, "div">((props, ref) => {
 
   return (
     <InputGroup
+      ref={ref}
       variant="outline"
       opacity={0.9}
       borderRadius="4px"
       size="md"
-      ref={ref}
+      boxShadow={`inset 0 0 0.5px 1px hsla(0, 0%,  
+        100%, 0.075),
+        /* shadow ring 👇 */
+        0 0 0 5px hsla(0, 0%, 0%, 0.05),
+        /* multiple soft shadows 👇 */
+        0 0.3px 0.4px hsla(0, 0%, 0%, 0.02),
+        0 0.9px 1.5px hsla(0, 0%, 0%, 0.045),
+        0 3.5px 6px hsla(0, 0%, 0%, 0.09);`}
       {...inputGroupProps}
     >
       <InputLeftElement>
@@ -108,17 +112,12 @@ const Adder = forwardRef<Props, "div">((props, ref) => {
         value={text}
         onChange={handleChange}
         onBlur={onAdd}
-        placeholder={mode} // it doesn't work if you pass it to input group for some reason
         onKeyDown={R.when((e) => e.key === "Enter", onAdd)}
         bg="gray.900"
+        placeholder={props.placeholder}
       />
     </InputGroup>
   );
 });
 
 export default Adder;
-
-export function getRandomEmoji(seed = "") {
-  const seededRandom = new Rand(seed);
-  return emojis[Math.floor(seededRandom.next() * emojis.length)];
-}
