+++ src/emojis.tsx
@@ -1,9 +1,13 @@
+import Rand from "rand-seed";
+import { emojis } from "./emojis";
+
+import { emojis } from "./emojis";
+
 export const emojis = [
   "😄",
   "😃",
   "😀",
   "😊",
-  "☺",
   "😉",
   "😍",
   "😘",
@@ -14,8 +18,6 @@ export const emojis = [
   "😝",
   "😛",
   "😳",
-  "😁",
-  "😔",
   "😌",
   "😒",
   "😞",
@@ -24,13 +26,11 @@ export const emojis = [
   "😂",
   "😭",
   "😪",
-  "😥",
   "😰",
   "😅",
   "😓",
   "😩",
   "😫",
-  "😨",
   "😱",
   "😠",
   "😡",
@@ -41,9 +41,6 @@ export const emojis = [
   "😷",
   "😎",
   "😴",
-  "😵",
-  "😲",
-  "😟",
   "😦",
   "😧",
   "😈",
@@ -63,8 +60,6 @@ export const emojis = [
   "💂",
   "👶",
   "👦",
-  "👧",
-  "👨",
   "👩",
   "👴",
   "👵",
@@ -72,10 +67,7 @@ export const emojis = [
   "👼",
   "👸",
   "😺",
-  "😸",
   "😻",
-  "😽",
-  "😼",
   "🙀",
   "😿",
   "😹",
@@ -131,7 +123,6 @@ export const emojis = [
   "💏",
   "💑",
   "👯",
-  "🙆",
   "🙅",
   "💁",
   "🙋",
@@ -624,41 +615,21 @@ export const emojis = [
   "🎭",
   "📍",
   "🚩",
-  "⬆",
-  "⬇",
-  "⬅",
-  "➡",
   "🔠",
   "🔡",
   "🔤",
-  "↗",
-  "↖",
-  "↘",
-  "↙",
-  "↔",
-  "↕",
   "🔄",
-  "◀",
-  "▶",
   "🔼",
   "🔽",
-  "↩",
-  "↪",
-  "ℹ",
   "⏪",
   "⏩",
   "⏫",
   "⏬",
-  "⤵",
-  "⤴",
   "🆗",
   "🔀",
   "🔁",
   "🔂",
-  "🆕",
   "🆙",
-  "🆒",
-  "🆓",
   "🆖",
   "📶",
   "🎦",
@@ -683,17 +654,11 @@ export const emojis = [
   "🅿",
   "♿",
   "🚭",
-  "🈷",
-  "🈸",
-  "🈂",
-  "Ⓜ",
   "🛂",
   "🛄",
   "🛅",
   "🛃",
   "🉑",
-  "㊙",
-  "㊗",
   "🆑",
   "🆘",
   "🆔",
@@ -706,11 +671,8 @@ export const emojis = [
   "🚷",
   "🚸",
   "⛔",
-  "✳",
-  "❇",
   "❎",
   "✅",
-  "✴",
   "💟",
   "🆚",
   "📳",
@@ -721,7 +683,6 @@ export const emojis = [
   "🅾",
   "💠",
   "➿",
-  "♻",
   "♈",
   "♉",
   "♊",
@@ -738,13 +699,6 @@ export const emojis = [
   "🔯",
   "🏧",
   "💹",
-  "💲",
-  "💱",
-  "©",
-  "®",
-  "™",
-  "〽",
-  "〰",
   "🔝",
   "🔚",
   "🔙",
@@ -757,64 +711,22 @@ export const emojis = [
   "❕",
   "❔",
   "🔃",
-  "🕛",
-  "🕧",
-  "🕐",
-  "🕜",
-  "🕑",
-  "🕝",
-  "🕒",
-  "🕞",
-  "🕓",
-  "🕟",
-  "🕔",
-  "🕠",
-  "🕕",
-  "🕖",
-  "🕗",
-  "🕘",
-  "🕙",
-  "🕚",
-  "🕡",
-  "🕢",
-  "🕣",
-  "🕤",
-  "🕥",
   "🕦",
-  "✖",
-  "➕",
-  "➖",
-  "➗",
-  "♠",
-  "♥",
-  "♣",
-  "♦",
   "💮",
   "💯",
-  "✔",
-  "☑",
   "🔘",
   "🔗",
   "➰",
   "🔱",
   "🔲",
-  "🔳",
-  "◼",
-  "◻",
-  "◾",
-  "◽",
-  "▪",
-  "▫",
   "🔺",
-  "⬜",
-  "⬛",
-  "⚫",
-  "⚪",
-  "🔴",
-  "🔵",
   "🔻",
   "🔶",
   "🔷",
   "🔸",
   "🔹",
 ];
+export function getRandomEmoji(seed = "") {
+  const seededRandom = new Rand(seed);
+  return emojis[Math.floor(seededRandom.next() * emojis.length)];
+}
